Rename a component, preparing for extension
