+++ .storybook/preview.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { Preview } from '@storybook/react'
 import { INITIAL_VIEWPORTS } from '@storybook/addon-viewport'
 import '../src/locales/i18n'
-import { withDefaultTheme } from '../src/app/components/ThemeByNetwork'
+import { withDefaultTheme } from '../src/app/components/ThemeByScope'
 import { initialize, mswLoader } from 'msw-storybook-addon'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import { handlers } from '../internals/mocks/msw-handlers'
